requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
